Refactor mutable Foundation variants out to their own files
